extract crab module
